requirements.txt
box-sdk-gen==1.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
docx-mailmerge==0.5.0
idna==3.10
iniconfig==2.0.0
lxml==5.3.0
numpy==2.1.1
packaging==24.1
pandas==2.2.2
pluggy==1.5.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-toolbelt==1.0.0
ruff==0.6.5
six==1.16.0
tzdata==2024.1
urllib3==2.2.3
